# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mon API Blog",
        "description": "Cette API permet de gérer un blog.",
        "contact": {
            "name": "Joséphin Sylvère",
            "email": "josephinsylvere@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/comments": {
            "post": {
                "tags": [
                    "Commentaires"
                ],
                "summary": "Crée un nouveau commentaire",
                "description": "Enregistre un nouveau commentaire avec les données fournies après validation.",
                "operationId": "createComment",
                "requestBody": {
                    "description": "Données du nouveau commentaire",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "content",
                                    "post_id",
                                    "user_id"
                                ],
                                "properties": {
                                    "content": {
                                        "description": "Contenu du commentaire",
                                        "type": "string"
                                    },
                                    "post_id": {
                                        "description": "ID du post associé",
                                        "type": "integer"
                                    },
                                    "user_id": {
                                        "description": "ID de l'utilisateur",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Commentaire créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Comment"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
        "/api/posts/{post_id}/comments": {
            "get": {
                "tags": [
                    "Commentaires"
                ],
                "summary": "Récupère tous les commentaires d'un post",
                "description": "Retourne tous les commentaires associés à un post spécifique.",
                "operationId": "getAllPostComments",
                "parameters": [
                    {
                        "name": "post_id",
                        "in": "path",
                        "description": "ID du post",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Opération réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Comment"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
        "/api/comments/{id}": {
            "get": {
                "tags": [
                    "Commentaires"
                ],
                "summary": "Récupère un commentaire par son ID",
                "description": "Retourne un commentaire spécifique par son ID.",
                "operationId": "getCommentById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du commentaire",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Opération réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Comment"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            },
            "put": {
                "tags": [
                    "Commentaires"
                ],
                "summary": "Met à jour un commentaire existant",
                "description": "Met à jour un commentaire avec les données fournies après validation.",
                "operationId": "updateComment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du commentaire à mettre à jour",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Données mises à jour du commentaire",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "content",
                                    "post_id",
                                    "user_id"
                                ],
                                "properties": {
                                    "content": {
                                        "description": "Contenu du commentaire",
                                        "type": "string"
                                    },
                                    "post_id": {
                                        "description": "ID du post associé",
                                        "type": "integer"
                                    },
                                    "user_id": {
                                        "description": "ID de l'utilisateur",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Commentaire mis à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Comment"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Commentaires"
                ],
                "summary": "Supprime un commentaire",
                "description": "Supprime un commentaire spécifié par son ID et renvoie un message de succès.",
                "operationId": "deleteComment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du commentaire à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Commentaire supprimé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Le commentaire a été supprimé avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Comment": {
                "title": "Comment",
                "description": "Un modèle de commentaire pour le blog",
                "properties": {
                    "id": {
                        "description": "L'identifiant unique du commentaire",
                        "type": "integer"
                    },
                    "post_id": {
                        "description": "L'identifiant du post associé",
                        "type": "integer"
                    },
                    "user_id": {
                        "description": "L'identifiant de l'utilisateur qui a écrit le commentaire",
                        "type": "integer"
                    },
                    "content": {
                        "description": "Le contenu du commentaire",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "La date et l'heure de création du commentaire",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "La date et l'heure de la dernière mise à jour du commentaire",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "auth_bearer": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Entrez le jeton JWT de cette manière: Bearer <token>"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}